feat(break): xml arrays now default to "un-wrapped" style
The global style can be set by xml.wrapped and xml.wrapped_arrays.
Right now xml.wrapped only affects the style of arrays, but in the
future this could be the global switch for other types as well.
For individual types the style can be set on the type definition
using xml.wrapped as before.