fix: type check was base on the name only
we have to consider the properties as well. nested types can lead to duplicated type-names but with different properties. right now the comparison skips a few properties, like the 'default' values so that we do not generate too many unnecessary types. this might need to be reconsidered in the future.

the second change in the 'sort_type' function makes sure, that we do not throw out, our newly created types in the sorting process. before the change, the types might still have been kept, but the fields could/would have pointed to the wrong name.